requirements.txt
redis
pyyaml
pymongo
tornado
scipy
numpy
